refactor[BE]: Refactor src dir tree
