+++ backend/src/api/auth/auth.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { AuthService } from './auth.service';
+import { PrismaModule } from 'src/prisma/prisma.module';
+import { JwtService } from '@nestjs/jwt';
+
+@Module({
+	imports: [
+		PrismaModule,
+	],
+	providers: [AuthService, JwtService],
+	exports: [AuthService],
+})
+export class AuthModule {}
