+++ backend/src/api/users/users.module.ts
@@ -1,18 +1,17 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { PrismaModule } from '../prisma/prisma.module';
+import { PrismaModule } from '../../prisma/prisma.module';
 import { UsersController } from './users.controller';
-import { MyController } from './my.controller';
 
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
-import { AuthService } from './auth/auth.service';
-import { FTStrategy } from './auth/ft.strategy';
+import { AuthService } from '../auth/auth.service';
+import { FTStrategy } from '../auth/ft.strategy';
 
 @Module({
   imports: [
     PrismaModule,
-    PassportModule,
+    PassportModule, 
     JwtModule.registerAsync({
         useFactory: () => ({
           secret: process.env.JWT_SECRET,
@@ -27,7 +26,6 @@ import { FTStrategy } from './auth/ft.strategy';
   ],
   controllers: [
     UsersController,
-    MyController
   ],
   exports: [UsersService],
 })
