+++ backend/src/app.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
-import { UsersModule } from './users/users.module';
 import { ConfigModule } from '@nestjs/config';
+import { UsersModule } from './api/users/users.module';
 import { PrismaModule } from './prisma/prisma.module';
+import { MyModule } from './api/my/my.module';
 
 @Module({
   imports: [
@@ -9,6 +10,7 @@ import { PrismaModule } from './prisma/prisma.module';
       isGlobal: true,
     }),
     UsersModule, 
+    MyModule,
     PrismaModule,
   ],
   controllers: [],
